Move a URLRequest extension into its own file
