+++ Sources/StubbornNetwork/RequestStub.swift
@@ -101,7 +101,7 @@ struct RequestStub: CustomDebugStringConvertible, Codable {
 extension RequestStub: Equatable {
     static func == (lhs: RequestStub, rhs: RequestStub) -> Bool {
         return lhs.data == rhs.data &&
-        lhs.request.matches(request: rhs.request) &&
+        lhs.request.matches(otherRequest: rhs.request) &&
             lhs.error?.localizedDescription == rhs.error?.localizedDescription &&
             lhs.response == rhs.response
     }
